requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
defusedxml==0.7.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
